+++ changelog.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog][keep-a-changelog]
 - Added Freja eID Organizational IDs
 - Changed how to decide module: require('eid-provider')('yourmodule')
 - Added involuntary dependency on request for bankid due to soap module. Hope this goes away.
+- Restructured documentation and added a bit of usefull stuff to it
+- Moved the location of modules into the modules sub-folder
 
 ## [0.0.2] (2020-06-06)
 - Added BankID via Svensk e-Identitet (gbankid)
@@ -21,5 +23,6 @@ The format is based on [Keep a Changelog][keep-a-changelog]
 
 [keep-a-changelog]: http://keepachangelog.com/en/1.0.0/
 [Unreleased]: https://github.com/DSorlov/eid-provider/compare/master...dev
+[0.0.3]: https://github.com/DSorlov/eid-provider/releases/tag/v0.0.3
 [0.0.2]: https://github.com/DSorlov/eid-provider/releases/tag/v0.0.2
 [0.0.1]: https://github.com/DSorlov/eid-provider/releases/tag/v0.0.1